去掉中英文前后的引号，修改文件结构，修复已知bug
