+++ server/db.js
@@ -37,4 +37,45 @@ trans.find(function (err, list) {
     }
 
 })
+var fs = require('fs')
+var path = require('path')
+function readFile() {
+    var cnList = []
+    var enList = []
+    var cnRes = fs.readFileSync(path.join(__dirname, '../../Docs/lang.cn.js'))
+    cnList = cnList.concat(cnRes.toString('utf-8', 13, cnRes.length - 28).split(',\r'))
+
+
+    var enRes = fs.readFileSync(path.join(__dirname, '../../Docs/lang.en.js'))
+    enList = enList.concat(enRes.toString('utf-8', 13, enRes.length - 28).split(',\r'))
+    enList.forEach((item, idx) => {
+        var identifer = item.split(':')[0]
+        var tempeName = item.split(':')[1].trim()
+        var eName = tempeName.slice(1, tempeName.length - 1)
+        var name = cnList[idx].split(':')[1].trim()
+        if (identifer == cnList[idx].split(':')[0]) {
+            //英文和中文索引一样的情况
+            trans.create({
+                identifer: identifer.trim(),
+                name: name.slice(1, name.length - 1),
+                eName,
+                location: 'Docs',
+                module: 'home',
+                branch: 'v1.0'
+            })
+        } else if (cnList.some(unit => unit.split(':')[0] == identifer)) {
+            //英文和中文索引不一样的情况
+            let tempName = cnList.find(unit => unit.split(':')[0] == identifer).split(':')[1].trim();
+            trans.create({
+                identifer: identifer.trim(),
+                name: tempName.slice(1, tempName.length - 1),
+                eName,
+                location: 'Docs',
+                module: 'home',
+                branch: 'v1.0'
+            })
+        }
+
+    })
+}
 module.exports = trans
